backend/requirements.txt
uvicorn
fastapi
happybase
pydantic
pydantic_settings
python-dotenv
